backend/requirements.txt
fastapi
uvicorn
pymongo
python-docx
jinja2
pdfkit
pydantic
python-multipart 